+++ package.json
@@ -82,7 +82,7 @@
     "concurrently": "7.2.2",
     "eslint": "8.18.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-deprecation": "1.3.2",
+    "eslint-plugin-deprecation": "^1.3.2",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-storybook": "0.5.13",
     "http-server": "14.1.1",
