--- src/ProofOfConcepts/DeleteModals/DeleteKafkaInstance/index.ts
@@ -1,1 +1,0 @@
-export * from "./DeleteKafkaInstance";
