--- src/ProofOfConcepts/DeleteModals/DeleteKafkaTopic/index.ts
@@ -1,1 +1,0 @@
-export * from "./DeleteKafkaTopic";
